chore: Organise cfd into a module
I am still new to rust and I used to have far more structure in my code. This is not a necessity for the tournament, but I was interested how this could be done in rust and came up with this. Happy to revert that if it's not rust idiomatic or objects the teams coding style.

Principle idea is to have all cfd related data and functions collected under the cfd namespace.

 - cfd::services: all cfd services (free floating functions)
 - cfd::models: all cfd models
 - cfd::dal: the data access layer. Note typically this layer is exposed through the service layer, but sometimes its just convenient to access the dal directly, like list_cfds.
