+++ rust/src/cfd/mod.rs
@@ -1,0 +1,17 @@
+pub mod models;
+mod open;
+mod settle;
+
+pub use dal::load_cfds;
+pub use open::open;
+pub use settle::settle;
+
+mod dal {
+    mod insert_cfd;
+    mod load_cfds;
+    mod update_cfd;
+
+    pub use insert_cfd::insert_cfd;
+    pub use load_cfds::load_cfds;
+    pub use update_cfd::update_cfd;
+}
